rename libraries' index files to index.js
