+++ libraries/app/package.json
@@ -24,5 +24,5 @@
     "promise": "^8.0.1",
     "uuid": "^3.3.2"
   },
-  "main": "./src/app.js"
+  "main": "./src/index.js"
 }
