+++ libraries/config/package.json
@@ -5,7 +5,7 @@
   "author": "Jonas Finnemann Jensen <jopsen@gmail.com>",
   "description": "Configuration loader that injects env variables into YAML",
   "license": "MPL-2.0",
-  "main": "./src/config",
+  "main": "./src/index.js",
   "scripts": {
     "lint": "eslint src/*.js test/*.js",
     "test": "mocha test/*_test.js"
