+++ libraries/config/src/index.js
